Split foldunfold/permissions.rs in two files
