+++ prusti-viper/src/encoder/foldunfold/mod.rs
@@ -7,7 +7,8 @@
 use self::path_ctxt::*;
 use crate::encoder::foldunfold::action::Action;
 use crate::encoder::foldunfold::perm::*;
-use crate::encoder::foldunfold::permissions::*;
+use crate::encoder::foldunfold::requirements::*;
+use crate::encoder::foldunfold::footprint::*;
 use crate::encoder::foldunfold::semantics::ApplyOnState;
 use crate::encoder::Encoder;
 use prusti_common::utils::to_string::ToString;
@@ -29,7 +30,8 @@ mod borrows;
 mod log;
 mod path_ctxt;
 mod perm;
-mod permissions;
+mod requirements;
+mod footprint;
 mod places_utils;
 mod semantics;
 mod state;
